Enable runtime loading of backends, renames
Backends now provide a generic Backend class for now this is just the
SPARQLHTTPBackend but will in the future enable QLever support.

Also don't use the name "sparql_backend" but the more generic "backend" to
refer to the loaded backend. On the other hand refer to the existing backend
as virtuoso. It might support other standard SPARQL HTTP endpoints too but
since as I understand this is largely untested make it more explicit.