QLever: still rough but can do queries
When trained with virtuoso this version can answer "Which character did Ellen
play in Finding Nemo?" using QLever as a backend.
This takes a very long time however which is due to several problems

0) Language in relation results in an explosion of the number of relations for
each entity as many relations are available in many languages
1) Missing support for count() in QLever means we currently receive the entire
result and count its size
2) Missing support for Keep-Alive connections in QLever results in a lot of
reconnects and this seems to be especially slow with urllib3's ConnectionPool.
At the minimum each TCP connect costs 3 network round trips

Also not working is support for OPTIONAL() so we might not get relations looked
up with 'include_name' which miss a name.