Fix evaluation bug
The conversion to python3's sorting concept with only keys
disabled a key = lambda x: x.query_candidate thus we were
getting the EvaluationCandidate instead of the QueryCandidate and
for that feature extraction doesn't work.

All in all the Python 3 sorting with keys only only works via an
ugly hack, it would be better to refactor so that QueryCandidate's
have comparison operators themselves even if they proxy to a ranker.