Add lax entity linking mode for empty case
When we fail to identify any entities, we would be unable to do anything
further for the query, however it may only be a POS tagging problem.
For example in "what did chadwick identify" chadwick was tagged as VB if not
capitalized. This change unbreaks this query and potentially many more.

Negative impact should be small since it would only hurt queries which
correctly contain no entities which doesn't make much sense.