Try a more symmetric Siamese CNN
Instead of only doing convolutions on the question also convolve the relations.
We also get rid of the dense_r_input layer. This is slightly better when
trained on WQ 1of2 and tested on WQ 2of2 and slightly worse when trained on all
of WQ train and tested on WQ test.

On 2of2 it looks like it generalizes somewhat better giving better top-5 scores
with all wrong-when-compared-to-old queries slipping no worse than 2/3 place.