Increase RelationNgramScorer regularization on WQ
For example in the query "who invented python" the relation_score is
unnaturally strong (0.08 of 0-1.0) for the wrong

"symbols.namesake.named_after"

relation.

It looks like this is due to that relation coming up with correct result sets
for multiple "invented" queries in the training set while the correct

"computer.programming_language.language_designers"

relation is never seen and lands on 4th place. More regularization pushes the
correct relation to the second place at least.