Further NN changes, use one embedding mat for all
This uses the same embedding for question and relation. Also it changes the way
we encode the relation, instead of averaging relation parts we simply take all
relation words. If the relation name is too long we take a suffix because often
there lie the more important terms.

This gives better results on the WQ test (51.3) than the previous change and
shows better generalization than the older version, in particular it allows
for example the query "what's the altitude of the ISS" to work even though
it's relation isn't in the training data. This was broken with the old
DeepRelScorer