answer_type_identifier: support new format
Natalie changed her format to include the original text and I changed
the freebaseized format to only include the mid and original text.

Currently this has terrible RAM use because now the answer_type_identifier
needs the whole entity -> type mapping in RAM but we already solved that
for the rest of Aqqu using RocksDB so in the next step we will use it here too.