Try out contrastive loss with a margin
We use a contrstive loss, that is instead of penalizing all differences between
the cosine similarity and the label we penalize the square of the difference
for positive examples and the square of the difference to 0 (uncorrelated) for
negative examples. Additionally for negative examples we allow for a margin
around 0 cosine similarity where we don't penalize. This should avoid changing
weights for more candidates that are judged as pretty uncorrelated already.

Looking at the loss in tensorboard it seems to go down more smoothly and with
somewhat less noise. Running on WebQSP the F1 score is unchanged but oracle 3
and oracle 5 increase slightly