Fix continuing training, naming
While we did load a model and continued training,
global_variables_initializer() actually reset all the weights.

Instead we only need to initialize the variables of the optimizer and
global_step. Maybe in the future we want to save these as well.

Sadly this is a bit tricky/ugly see: https://stackoverflow.com/q/35164529