Make Backend reuse the learner container
This will not only simplify things in the Dockerfile it also means we are
guaranteed to use the exact code of the learner by the same name. However it
means you MUST use the same name for the backend as the learner. May add a way
to change the instance name