Store models in extra <name>_models_vol volume
Instead of storing models with other data that is shared between Aqqu
instances, store them in an extra volume named by the build_and_run.sh script.
This makes models of different learning runs independent for arbitrary changes.

This clashes somewhat with the original concept of named rankers but that only
allows for certrain changes and has been quite cumbersome to deal with.
Longterm it should be completely phased out.

Note that the debug mode does not replace the models!