Add preliminary support for SimpleQuestions
in its current form we learn and evaluate on SimpleQuestions with the same
Precision/Recall/F1 metric as on the other datasets. While it works ok this
gives weird results and is likely less effective.

This is due to the fact that SimpleQuestions only picks a single target entity
even for questions which would correctly be answered with a set. Normally SQ is
therefore trained and evaluated with matching the fact relation.

Still some changes are needed to support using MIDs instead of the names for
training, however we can reuse this for WebQSP