Allow overriding ranker parameters on the cli
This needs quite a few changes to the scorer list/map, learning functions and
ranker. Also we update the documentation and build_and_run.sh script.
Additionally we can now use this for hyperparameter optimization in RiseML.

For this CV learning is used, which entails some fixes since this has been
untested for a while.