Rework scorer configuration everything is in Conf
Instead of having the defaults as default parameters in DeepCNNAqquRelScorer
and AqquModel we now use static 'default_config' properties that get loaded by
scorer_globals.Conf. Thus we not only see the overriden configuration
parameters but everything. This is true for the config written in learner.cv()
as well as for the '/config' path of the translation server.